Settings CRUD
Patch settings page to work with the new setup
